requirements.txt
discord.py>=2.3.0
python-dotenv>=0.19.0
asyncio-cron>=0.0.2